Item6578: Fix for: Use of uninitialized value in split at lib/TWiki.pm line 1302
git-svn-id: http://develop.twiki.org/svn/twiki/trunk@19445 a00a5322-12db-0310-a70b-8735589c885e

+++ core/lib/TWiki.pm
@@ -1300,10 +1300,13 @@ sub cacheIconData {
 
         } else {
             # Remove icon cache if any topic in the ICONTOPIC list changed
-            foreach my $iconTopic (split(/ *, */, $this->{prefs}->getPreferencesValue( 'ICONTOPIC' ))) {
-                my( $iWeb, $iTopic ) = $this->normalizeWebTopicName( $this->{webName}, $iconTopic );
-                if( ( $web eq $iWeb ) && ( $topic eq $iTopic ) ) {
-                    unlink( $cacheFile );
+            my $topics = $this->{prefs}->getPreferencesValue( 'ICONTOPIC' );
+            if( $topics ) {
+                foreach my $iconTopic ( split( / *, */, $topics ) ) {
+                    my( $iWeb, $iTopic ) = $this->normalizeWebTopicName( $this->{webName}, $iconTopic );
+                    if( ( $web eq $iWeb ) && ( $topic eq $iTopic ) ) {
+                        unlink( $cacheFile );
+                    }
                 }
             }
         }
